backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
langchain
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
langchain-core